swap out mocha/chai for ava
