+++ package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "lint": "eslint src",
-    "test": "ava ava/*.test.js",
+    "test": "ava test/*.test.js --serial",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- test/**/*.test.js --reporter spec",
     "test-travis": "npm run test-cov"
   },
@@ -28,14 +28,10 @@
   "license": "MIT",
   "devDependencies": {
     "ava": "^0.15.2",
-    "chai": "^3.4.1",
-    "chai-as-promised": "^5.3.0",
     "coveralls": "^2.11.4",
     "eslint": "^3.2.0",
     "eslint-config-airbnb-base": "^5.0.1",
     "eslint-plugin-import": "^1.12.0",
-    "istanbul": "^0.4.0",
-    "mocha": "^2.3.3",
     "mongodb": "^2.0.48",
     "mongoose": "^4.5.5"
   }
